proj/requirements.txt
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
fonttools==4.39.3
jdatetime==4.1.1
openpyxl==3.1.2
requests==2.28.2